+++ src/include/embedded/repeat_call.hpp
@@ -1,0 +1,12 @@
+﻿// -*- coding: utf-8-with-signature-dos -*-
+// vim:fileencoding=utf-8:ff=dos
+//
+// Repeat call function utility.
+
+#ifndef CUN_REPEAT_CALL_HPP_INCLUDED
+#define CUN_REPEAT_CALL_HPP_INCLUDED
+
+// For this library
+#include "repeat_call_core.hpp"
+
+#endif // ndef CUN_REPEAT_CALL_HPP_INCLUDED